requirements.txt
dash==2.17.1
plotly==5.19.0
pandas==2.2.3
requests==2.32.3
schedule==1.2.2
python-dotenv==1.0.1
gunicorn==22.0.0
dash-table==5.0.0